pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "silver-layer"
version = "2.0.0"
description = "Config-driven Bronze → Silver pipeline for Databricks (transformation, DQX, Unity Catalog governance)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [{ name = "Silver Layer Team" }]
keywords = ["databricks", "delta", "etl", "silver-layer", "data-quality", "unity-catalog"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyyaml>=6.0",
    "pyspark>=3.3.0",
    "delta-spark>=2.3.0",
]

[project.optional-dependencies]
databricks = [
    "databricks-labs-dqx",
    "databricks-sdk",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]

[tool.setuptools.packages.find]
where = ["."]
include = ["utils*"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
